requirements.txt
Flask>=0.10.1
flask-mongoengine>=0.7.1
Flask-WTF>=0.11
mongoengine>=0.8.7
numpy>=1.6.2
Werkzeug>=0.10.1
WTForms>=2.0.2
ExifRead>=1.4.2
uwsgi>=2.0.3
Pillow>=2.7.0
Flask-RESTful>=0.3.2